% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-space.R
\name{ets_label}
\alias{ets_label}
\title{Label of an ETS model (e.g. "AAdN")}
\usage{
ets_label(model_components)
}
\arguments{
\item{model_components}{Vector (error, trend, season, damped).}
}
\value{
Character string.
}
\description{
Label of an ETS model (e.g. "AAdN")
}
\keyword{internal}
