% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simulate_future_trajectories}
\alias{simulate_future_trajectories}
\title{Simulate Future Trajectories for an ETS model}
\usage{
simulate_future_trajectories(bets_fit, h = 10, n_traj = 1000)
}
\arguments{
\item{bets_fit}{A BETS fit object.}

\item{h}{The forecast horizon.}

\item{n_traj}{The number of trajectories to simulate.}
}
\value{
A matrix of future trajectories.
}
\description{
Simulate Future Trajectories for an ETS model
}
\keyword{internal}
