% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesETS-package.R
\docType{package}
\name{BayesETS-package}
\alias{BayesETS}
\alias{BayesETS-package}
\title{BayesETS: Forecasting using Bayesian ETS}
\description{
Bayesian exponential smoothing (ETS) state space models for univariate time series, accounting for both parameter and model uncertainty. The initial states and the error variance are integrated out analytically, under a conjugate Normal-inverse-chi-squared prior; the smoothing parameters are integrated numerically, by adaptive Gauss-Hermite quadrature (AGHQ) or adaptive importance sampling (AIS). The ETS models are combined by Bayesian model averaging (BMA), and forecasts are obtained by simulating future sample paths. Currently only additive models are supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LorenzoZambon/BETS}
  \item Report bugs at \url{https://github.com/LorenzoZambon/BETS/issues}
}

}
\author{
\strong{Maintainer}: Lorenzo Zambon \email{lorenzo.zambon@supsi.ch}

Authors:
\itemize{
  \item Lorenzo Zambon \email{lorenzo.zambon@supsi.ch}
}

}
\keyword{internal}
