% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BETS-package}
\alias{BETS}
\alias{BETS-package}
\title{BETS package}
\description{
Bayesian ETS models.
}
\author{
\strong{Maintainer}: Lorenzo Zambon \email{lorenzo.zambon@supsi.ch}

Authors:
\itemize{
  \item Lorenzo Zambon \email{lorenzo.zambon@supsi.ch}
}

}
\keyword{internal}
