% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_stacking_weights}
\alias{compute_stacking_weights}
\title{Stacking weights}
\usage{
compute_stacking_weights(log_lik_list)
}
\arguments{
\item{log_lik_list}{List with one matrix of pointwise log-likelihoods
(draws x time) per model.}
}
\value{
Vector of stacking weights.
}
\description{
Stacking weights
}
\keyword{internal}
