% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_stacking_weights}
\alias{compute_stacking_weights}
\title{Compute Stacking Weights}
\usage{
compute_stacking_weights(log_lik_list)
}
\arguments{
\item{log_lik_list}{A list of matrices containing pointwise log-likelihoods for each model.}
}
\value{
A vector of stacking weights.
}
\description{
Compute Stacking Weights
}
\keyword{internal}
