% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ets_label}
\alias{ets_label}
\title{Label ETS model components}
\usage{
ets_label(model_components)
}
\arguments{
\item{model_components}{A character vector of ETS model components.}
}
\value{
A string representing the ETS model label.
}
\description{
Label ETS model components
}
\keyword{internal}
