% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bets.R
\name{forecast.bets}
\alias{forecast.bets}
\title{Forecast from a BETS model}
\usage{
\method{forecast}{bets}(object, h = 10, level = c(80, 95), n_traj = NULL, ...)
}
\arguments{
\item{object}{A fitted object from \code{\link[=bets]{bets()}}.}

\item{h}{Forecast horizon.}

\item{level}{Confidence levels for intervals.}

\item{n_traj}{Number of simulated trajectories.}

\item{...}{Unused.}
}
\value{
An object with class \code{c("bets_forecast", "forecast")}.
}
\description{
Produces forecast distributions from a fitted \code{bets} object using
Monte Carlo trajectories.
}
