% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{simulate_future_trajectories}
\alias{simulate_future_trajectories}
\title{Future trajectories of the model combination}
\usage{
simulate_future_trajectories(bets_fit, h = 10, n_traj = 1000)
}
\arguments{
\item{bets_fit}{The \code{fit} element of a \code{bets} object.}

\item{h}{Forecast horizon.}

\item{n_traj}{Number of trajectories.}
}
\value{
Matrix of trajectories (n_traj x h).
}
\description{
Each model contributes a number of trajectories proportional to its weight.
}
\keyword{internal}
